+++ src/main/java/xyz/nikitacartes/easyauth/EasyAuth.java
@@ -25,6 +25,7 @@ import java.util.concurrent.TimeUnit;
 
 import static xyz.nikitacartes.easyauth.config.ConfigMigration.*;
 import static xyz.nikitacartes.easyauth.config.MainConfigV1.CURRENT_CONFIG_VERSION;
+import static xyz.nikitacartes.easyauth.config.StorageConfigV1.getDbApi;
 import static xyz.nikitacartes.easyauth.utils.EasyLogger.*;
 
 public class EasyAuth {
@@ -46,13 +47,7 @@ public class EasyAuth {
 
 
     public static void loadDatabase() {
-        if (storageConfig.databaseType.equalsIgnoreCase("mysql")) {
-            DB = new MySQL(storageConfig);
-        } else if (storageConfig.databaseType.equalsIgnoreCase("mongodb")) {
-            DB = new MongoDB(storageConfig);
-        } else {
-            DB = new SQLite(storageConfig);
-        }
+        DB = getDbApi();
         try {
             DB.connect();
         } catch (DBApiException e) {